histo-preproc/requirements.txt
numpy
openslide-python
SimpleITK
argparse
pyvips
